requirement.txt
mlflow>=2.8.0
cryptography==41.0.5
evidently==0.4.0
pandas>=1.5.0
numpy==1.24.3
scikit-learn>=1.2.0
seaborn>=0.12.0
matplotlib>=3.6.0
deepchecks==0.17.7
scipy>=1.10.0
PyYAML>=6.0
fastapi>=0.104.0
uvicorn>=0.24.0
pytest>=7.4.0
httpx>=0.25.0
pydantic>=2.0.0
